pyproject.toml
[project]
name = "arxiv-mcp-aj"
version = "0.1.4"
description = "Model Context Protocol server for arXiv PDF retrieval and LLM context generation."
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [{name = "arxiv MCP Team"}]
dependencies = [
  "mcp>=1.0.0",
  "arxiv>=2.1.0",
  "httpx>=0.27.0",
  "pymupdf>=1.24.0",
  "pydantic>=2.0.0",
  "tenacity>=8.2.0",
  "tiktoken>=0.7.0",
  "structlog>=24.0.0",
  "aiofiles>=23.0.0",
  "python-dotenv>=1.0.0",
  "anyio>=4.0.0"
]

[project.urls]
"Homepage" = "https://github.com/Aldrin-Joan/arXiv-mcp"
"Repository" = "https://github.com/Aldrin-Joan/arXiv-mcp"

[project.scripts]
arxiv-mcp = "src.console:entrypoint"

[tool.setuptools]
package-dir = {"" = ".", "src" = "src"}

[tool.setuptools.packages.find]
where = [".", "src"]
include = ["arxiv_mcp", "src", "src.*"]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"
